pyproject.toml
[project]
name = "jspf-converter"
version = "0.1.0"
description = "Flask playlist sync service for M3U, JSPF, ListenBrainz, and Octo-Fiesta handoff workflows."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "rapidfuzz<3.14",
    "waitress>=3.0.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=9.0.2",
    "ruff>=0.15.9",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-ra"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
